requirements.txt
Flask~=3.0.2
reportlab~=4.1.0
gunicorn==20.1.0
pytest~=8.1.1
pillow~=10.2.0
pandas~=2.2.1
pdf2docx~=0.5.8
Werkzeug~=3.0.1
pyClamd~=0.4.0
Werkzeug~=3.0.1
pyClamd~=0.4.0
python-json-logger~=2.0.7
